Update OrderController class

+++ src/Controller/OrderController.php
@@ -7,6 +7,7 @@ use Framework\DataMapper\CartMapper;
 use Framework\DataMapper\OrderStatusMapper;
 use Framework\DataMapper\UserMapper;
 use Framework\Core\Controller;
+use src\Model\Cart;
 
 class OrderController extends Controller
 {
@@ -15,10 +16,12 @@ class OrderController extends Controller
     private OrderStatusMapper $orderStatusMapper;
     private UserMapper $userMapper;
     private Auth $auth;
+    private Cart $cart;
 
     public function __construct(array $params)
     {
         parent::__construct();
+        $this->cart = new Cart();
         $this->cartMapper = new CartMapper();
         $this->auth = new Auth();
         $this->userMapper = new UserMapper();
@@ -28,12 +31,26 @@ class OrderController extends Controller
     public function orderAction(): void
     {
         if ($this->auth->isAuth()) {
-            if ($this->addOrder()) {
-                $productIdFromCart = $this->getOrderProductFromParams();
-                foreach ($productIdFromCart as $id) {
-                    $this->cartMapper->deleteProductFromCartByColumn(['product_id' => $id], $this->getUserId());
+            if ($this->params['button'] === 'Checkout' || $this->params['button'] === 'Buy') {
+                if ($this->addOrder()) {
+                    $productIdFromCart = $this->getOrderProductFromParams();
+                    foreach ($productIdFromCart as $id) {
+                        $this->cartMapper->deleteProductFromCartByColumn(['product_id' => $id], $this->getUserId());
+                    }
+                    header('Location:/cart');
                 }
-                header('Location:/cart');
+            }
+            if ($this->params['button'] === 'Add to cart') {
+                $cart = new Cart();
+                $cart->setUserId($this->getUserId());
+                $cart->setProductId((int)$this->params['product_id']);
+                if (isset($this->params['quantity'])) {
+                    $cart->setQuantity($this->params['quantity']);
+                } else {
+                    $cart->setQuantity(1);
+                }
+                $this->cartMapper->addProductsInCart($cart);
+                header('Location:/products');
             }
         } else {
             header('Location:/account/login');
@@ -58,7 +75,7 @@ class OrderController extends Controller
         $orderProductId = [];
         for ($i = 0; $i < count($this->params); $i++) {
             foreach ($this->params as $key => $param) {
-                if ($key == 'product_id_' . $i) {
+                if ($key == 'product_id_' . $i || $key == 'product_id') {
                     $orderProductId[] = $param;
                 }
             }
